requirements.txt
python-telegram-bot==20.8
asyncpg
python-dotenv
requests
typing_extensions
urllib3
apscheduler
qrcode
json